Seeders and migrations

+++ Backend/database/migrations/2023_11_30_101155_create_products_table.php
@@ -9,7 +9,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('products', function (Blueprint $table) {
-            $table->increments('ProductID');
+            $table->unsignedBigInteger('ProductID');
             $table->string('ProductName');
             $table->longText('Description');
             $table->double('Price');
